Fix jQuery types, TSConfig paths
