+++ .storybook/preview.js
@@ -1,15 +1,14 @@
 import Twig from 'twig';
 import { addDecorator } from '@storybook/react';
 import { useEffect } from '@storybook/client-api';
-import once from '@drupal/once';
 import twigDrupal from 'twig-drupal-filters';
 import twigAttributes from 'add-attributes-twig-extension';
 import keysort from '../lib/keysort';
 import uniqueId from '../lib/uniqueId';
+import './stubs/drupal';
+import './stubs/once';
 
 import '../dist/css/styles.css';
-import './_drupal';
-global.once = once;
 
 function setupTwig(twig) {
   twig.cache();
