+++ .storybook/stubs/drupalSettings.js
@@ -1,0 +1,3 @@
+window.drupalSettings = {};
+
+export default window.drupalSettings;
