+++ .storybook/stubs/once.js
@@ -1,0 +1,5 @@
+import once from '@drupal/once';
+
+window.once = once;
+
+export default once;
