+++ package.json
@@ -36,6 +36,7 @@
     "@storybook/react": "^6.4.14",
     "@storybook/storybook-deployer": "^2.8.10",
     "@storybook/theming": "^6.4.14",
+    "@types/jquery": "3.2",
     "@types/react": "^17.0.38",
     "@types/react-dom": "^17.0.11",
     "babel-loader": "^8.2.2",
