+++ source/03-components/back-to-top/back-to-top.es6.ts
@@ -1,3 +1,5 @@
+import Drupal from 'drupal';
+
 Drupal.behaviors.backToTop = {
   attach(context, settings) {
     const threshold = settings?.gesso?.backToTopThreshold ?? 200;
