+++ source/03-components/overlay-menu/overlay-menu.es6.ts
@@ -1,3 +1,4 @@
+import Drupal from 'drupal';
 import OverlayMenu from './modules/_OverlayMenu.es6';
 
 Drupal.behaviors.overlayMenu = {
