+++ tsconfig.json
@@ -13,10 +13,17 @@
     "typeRoots": [
       "node_modules/@types",
       "source/@types"
-    ]
+    ],
+    "baseUrl": ".",
+    "paths": {
+      "jquery": ["node_modules/jquery/dist/jquery.js"],
+      "drupal": [".storybook/stubs/drupal.js"],
+      "drupalSettings": [".storybook/stubs/drupalSettings.js"],
+      "once": [".storybook/stubs/once.js"]
+    }
   },
   "exclude": [
     "node_modules",
-    "lib/**",
+    "lib/**"
   ]
 }
